refactor: fix React hooks deps, extract modules, remove unused exports
Fix missing/incorrect useCallback and useEffect dependency arrays
across 15 components and hooks. Extract AppRouter from main.jsx,
useToast from Toast.tsx, and useSettingsLayout from SidebarModal.tsx
into standalone modules. Hoist render-phase side effects into
useLayoutEffect/useMemo, use lazy useState initializers, and remove
unused exports (badgeVariants, buttonVariants, toast helpers).
