+++ src/components/EmailVerificationStep.tsx
@@ -51,7 +51,7 @@ export default function EmailVerificationStep({ email, onVerified }: EmailVerifi
     return () => {
       if (pollRef.current) clearInterval(pollRef.current);
     };
-  }, [email, verified, onVerified]);
+  }, [email, verified, onVerified, t]);
 
   const handleResend = useCallback(async () => {
     if (resendCooldown > 0 || isResending || !OPENWHISPR_API_URL) return;
