+++ src/components/SettingsModal.tsx
@@ -11,7 +11,7 @@ import {
   Shield,
   MessageSquare,
 } from "lucide-react";
-import SidebarModal, { SidebarItem } from "./ui/SidebarModal";
+import SidebarModal, { type SidebarItem } from "./ui/SidebarModal";
 import SettingsPage, { SettingsSectionType } from "./SettingsPage";
 
 export type { SettingsSectionType };
