+++ src/components/ui/LanguageSelector.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect } from "react";
+import React, { useState, useRef, useEffect, useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import { createPortal } from "react-dom";
 import { ChevronDown, Search, X, Check } from "lucide-react";
@@ -38,7 +38,10 @@ export default function LanguageSelector({
   const [searchQuery, setSearchQuery] = useState("");
   const [highlightedIndex, setHighlightedIndex] = useState(0);
   const [dropdownPosition, setDropdownPosition] = useState({ top: 0, left: 0, width: 0 });
-  const containerRef = useRef<HTMLDivElement>(null);
+  const [portalTarget, setPortalTarget] = useState<HTMLElement | null>(
+    typeof document === "undefined" ? null : document.body
+  );
+  const containerRef = useRef<HTMLDivElement | null>(null);
   const dropdownRef = useRef<HTMLDivElement>(null);
   const triggerRef = useRef<HTMLButtonElement>(null);
   const searchInputRef = useRef<HTMLInputElement>(null);
@@ -51,25 +54,24 @@ export default function LanguageSelector({
       )
     : items;
 
-  useEffect(() => {
+  const handleSearchQueryChange = useCallback((value: string) => {
+    setSearchQuery(value);
     setHighlightedIndex(0);
-  }, [searchQuery]);
+  }, []);
 
   // Determine the portal container: use the closest dialog if inside one (to stay
   // within Radix's focus trap), otherwise fall back to document.body.
-  const portalTarget = useRef<HTMLElement>(document.body);
-
-  useEffect(() => {
-    if (containerRef.current) {
-      const dialog = containerRef.current.closest('[role="dialog"]');
-      portalTarget.current = (dialog as HTMLElement) ?? document.body;
-    }
+  const setContainerNode = useCallback((node: HTMLDivElement | null) => {
+    containerRef.current = node;
+    if (!node) return;
+    const dialog = node.closest('[role="dialog"]');
+    setPortalTarget((dialog as HTMLElement) ?? document.body);
   }, []);
 
   useEffect(() => {
-    if (isOpen && triggerRef.current) {
+    if (isOpen && triggerRef.current && portalTarget) {
       const triggerRect = triggerRef.current.getBoundingClientRect();
-      const target = portalTarget.current;
+      const target = portalTarget;
       // When portaled into a transformed ancestor (e.g. Radix Dialog),
       // fixed positioning is relative to that ancestor, not the viewport.
       const offsetX = target === document.body ? 0 : target.getBoundingClientRect().left;
@@ -83,7 +85,7 @@ export default function LanguageSelector({
         searchInputRef.current?.focus();
       });
     }
-  }, [isOpen]);
+  }, [isOpen, portalTarget]);
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
       const target = event.target as Node;
@@ -127,7 +129,7 @@ export default function LanguageSelector({
       case "Escape":
         e.preventDefault();
         setIsOpen(false);
-        setSearchQuery("");
+        handleSearchQueryChange("");
         break;
     }
   };
@@ -135,18 +137,18 @@ export default function LanguageSelector({
   const handleSelect = (languageValue: string) => {
     onChange(languageValue);
     setIsOpen(false);
-    setSearchQuery("");
+    handleSearchQueryChange("");
   };
 
   const clearSearch = () => {
-    setSearchQuery("");
+    handleSearchQueryChange("");
     if (searchInputRef.current) {
       searchInputRef.current.focus();
     }
   };
 
   return (
-    <div className={`relative ${className}`} ref={containerRef}>
+    <div className={`relative ${className}`} ref={setContainerNode}>
       {/* Trigger button - premium, tight, tactile macOS-style */}
       <button
         ref={triggerRef}
@@ -184,6 +186,7 @@ export default function LanguageSelector({
 
       {/* Dropdown - ultra-premium glassmorphic panel (rendered via portal) */}
       {isOpen &&
+        portalTarget &&
         createPortal(
           <div
             ref={dropdownRef}
@@ -203,7 +206,7 @@ export default function LanguageSelector({
                     ref={searchInputRef}
                     type="text"
                     value={searchQuery}
-                    onChange={(e) => setSearchQuery(e.target.value)}
+                    onChange={(e) => handleSearchQueryChange(e.target.value)}
                     onKeyDown={handleKeyDown}
                     placeholder={t("languageSelector.searchPlaceholder")}
                     className="w-full h-7 pl-7 pr-6 text-xs bg-transparent text-foreground border-0 focus:outline-none placeholder:text-muted-foreground/50"
@@ -265,7 +268,7 @@ export default function LanguageSelector({
               )}
             </div>
           </div>,
-          portalTarget.current
+          portalTarget
         )}
     </div>
   );
