+++ src/components/ui/Toast.tsx
@@ -1,32 +1,7 @@
 import * as React from "react";
 import { X, Copy, Check } from "lucide-react";
 import { cn } from "../lib/utils";
-
-export interface ToastProps {
-  id?: string;
-  title?: string;
-  description?: string;
-  action?: React.ReactNode;
-  variant?: "default" | "destructive" | "success";
-  duration?: number;
-  onClose?: () => void;
-}
-
-export interface ToastContextType {
-  toast: (props: Omit<ToastProps, "id">) => void;
-  dismiss: (id?: string) => void;
-  toastCount: number;
-}
-
-const ToastContext = React.createContext<ToastContextType | undefined>(undefined);
-
-export const useToast = () => {
-  const context = React.useContext(ToastContext);
-  if (!context) {
-    throw new Error("useToast must be used within a ToastProvider");
-  }
-  return context;
-};
+import { ToastContext, type ToastProps } from "./useToast";
 
 interface ToastState extends ToastProps {
   id: string;
@@ -312,19 +287,4 @@ const Toast: React.FC<
       )}
     </div>
   );
-};
-
-export const toast = {
-  success: (message: string) => ({
-    title: message,
-    variant: "success" as const,
-  }),
-  error: (message: string) => ({
-    title: message,
-    variant: "destructive" as const,
-  }),
-  info: (message: string) => ({
-    title: message,
-    variant: "default" as const,
-  }),
 };
