+++ src/components/ui/button.tsx
@@ -143,4 +143,4 @@ function Button({
   );
 }
 
-export { Button, buttonVariants };
+export { Button };
