+++ src/hooks/useModelDownload.ts
@@ -2,7 +2,7 @@ import { useState, useCallback, useEffect, useRef } from "react";
 import type { TFunction } from "i18next";
 import { useTranslation } from "react-i18next";
 import { useDialogs } from "./useDialogs";
-import { useToast } from "../components/ui/Toast";
+import { useToast } from "../components/ui/useToast";
 import type { WhisperDownloadProgressData } from "../types/electron";
 import "../types/electron";
