+++ src/hooks/useSettings.ts
@@ -73,6 +73,7 @@ export interface AgentModeSettings {
 
 function useSettingsInternal() {
   const store = useSettingsStore();
+  const { setCustomDictionary } = store;
 
   // One-time initialization: sync API keys, dictation key, activation mode,
   // UI language, and dictionary from the main process / SQLite.
@@ -94,11 +95,11 @@ function useSettingsInternal() {
     if (typeof window === "undefined" || !window.electronAPI?.onDictionaryUpdated) return;
     const unsubscribe = window.electronAPI.onDictionaryUpdated((words: string[]) => {
       if (Array.isArray(words)) {
-        store.setCustomDictionary(words);
+        setCustomDictionary(words);
       }
     });
     return unsubscribe;
-  }, [store.setCustomDictionary]);
+  }, [setCustomDictionary]);
 
   // Auto-learn corrections from user edits in external apps
   const [autoLearnCorrections, setAutoLearnCorrectionsRaw] = useLocalStorage(
